+++ store/store.go
@@ -32,12 +32,6 @@ const (
 	PingType   = "ping"
 )
 
-// 5 minutes
-const DefaultTTL = 60 * 5
-
-// 1 year
-const YearTTL = 60 * 24 * 365
-
 type S interface {
 	Push(key model.Key, item OwnableItem) error
 	Get(key model.Key) (OwnableItem, error)